Made Auto Align Subsystem
skibidi
